requirements.txt
youtube-transcript-api==0.6.2
streamlit
google-generativeai
python-dotenv
pathlib 
flask